% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_loss_model_compound_poisson.R
\name{compound_poisson}
\alias{compound_poisson}
\title{Estimate a compound Poisson distribution.}
\usage{
compound_poisson(poisson_mean, lambda, p, xmax, N = 2^12, tol = 0.005)
}
\arguments{
\item{poisson_mean}{Expected value of Poisson random variable.}

\item{lambda}{A vector of mle's of rate parameters from exponential mixture distribution.}

\item{p}{A vector of mle's of mixing proportions from exponential mixture distribution.}

\item{xmax}{A large integer representing a truncation point in the domain of exponential
mixture distribution.}

\item{N}{A large integer representing number of discretization points.}

\item{tol}{A number n such that P(Nmax > n) <= \code{tol} where Nmax has Poisson
with \code{possion_mean} distribution.}
}
\value{
A list containing the support \code{x} of the convolution distribution and
an estimate \code{density} of the compound Poisson distribution.
}
\description{
Estimate a compound Poisson distribution.
}
