% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censored_mixed_exponential_em.R
\name{em_censored_fit}
\alias{em_censored_fit}
\title{Compute the maximum likelihood estimates of an exponential mixture distribution and number of mixtures when
the data is right censored.}
\usage{
em_censored_fit(x, cens_point, n_c, prec = 1e-08, prec2 = 1e-04,
  itermax = 100, rmax = 10)
}
\arguments{
\item{x}{A data vector.}

\item{cens_point}{Point at which data is censored.}

\item{n_c}{Number of right censored data points.}

\item{prec}{Precision used for testing convergence of mle's as measured by
squared difference of loglikelihood functions.}

\item{prec2}{Estimated mixing proportions below \code{prec2} will cause algorithm to stop.}

\item{itermax}{Maximum number of EM algorithm iterations.}

\item{rmax}{Maximum number of mixtures to find.}
}
\value{
A list of mixing proportions \code{p} and associated means \code{mean}.
}
\description{
Compute the maximum likelihood estimates of an exponential mixture distribution and number of mixtures when
the data is right censored.
}
