% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_mixed_exponential_em.R
\name{em_group_fit}
\alias{em_group_fit}
\title{Compute the maximum likelihood estimates of an exponential mixture distribution and number of mixtures when
the data is grouped.}
\usage{
em_group_fit(x_grp, boundaries = NULL, data_type, prec = 1e-08,
  prec2 = 1e-04, itermax = 100, rmax = 10)
}
\arguments{
\item{x_grp}{Grouped data vector.}

\item{boundaries}{Interval boundaries of the grouped data.}

\item{data_type}{Three possible \code{data_type}s are \code{'interval', 'cross tab', or 'tabular'}}

\item{prec}{Precision used for testing convergence of mle's as measured by
squared difference of loglikelihood functions.}

\item{prec2}{Estimated mixing proportions below \code{prec2} will cause algorithm to stop.}

\item{itermax}{Maximum number of EM algorithm iterations.}

\item{rmax}{Maximum number of mixtures to find.}
}
\value{
A list of mixing proportions \code{p} and associated means \code{mean}.
}
\description{
Compute the maximum likelihood estimates of an exponential mixture distribution and number of mixtures when
the data is grouped.
}
